restructure shipment aggregate
