+++ shipment/event-handler/src/main/resources/application.yaml
