+++ shipment/query-side/build.gradle
