+++ shipment/query-side/src/main/java/org/example/event/sourcing/order/poc/query/shipment/QuerySideShipmentApplication.java
