+++ shipment/query-side/src/main/resources/application.yaml
